:sparkles: Refactored registry & registry keys & added Registries class
